Extract grpc code
