+++ crates/enclave/core/src/lib.rs
@@ -13,18 +13,11 @@
 )]
 
 use cosmrs::AccountId;
-use k256::ecdsa::VerifyingKey;
-use quartz_proto::quartz::{
-    core_server::Core, InstantiateRequest, InstantiateResponse, SessionCreateRequest,
-    SessionCreateResponse, SessionSetPubKeyRequest, SessionSetPubKeyResponse,
-};
 use serde::de::DeserializeOwned;
-use tonic::{Request, Response, Status};
 
 use crate::{
     attestor::{Attestor, DefaultAttestor},
     chain_client::{default::DefaultChainClient, ChainClient},
-    handler::Handler,
     key_manager::{default::DefaultKeyManager, KeyManager},
     kv_store::{default::DefaultKvStore, ConfigKey, ContractKey, NonceKey, TypedStore},
 };
@@ -32,6 +25,7 @@ use crate::{
 pub mod attestor;
 pub mod chain_client;
 pub mod error;
+pub mod grpc;
 pub mod handler;
 pub mod key_manager;
 pub mod kv_store;
@@ -95,35 +89,5 @@ where
 
     async fn store(&self) -> Self::Store {
         self.store.clone()
-    }
-}
-
-#[async_trait::async_trait]
-impl<A, C, K, S> Core for DefaultEnclave<A, C, K, S>
-where
-    A: Attestor + Clone,
-    C: ChainClient<Contract = AccountId> + Clone,
-    K: KeyManager<PubKey = VerifyingKey> + Clone,
-    S: TypedStore<ContractKey<AccountId>> + TypedStore<NonceKey> + TypedStore<ConfigKey> + Clone,
-{
-    async fn instantiate(
-        &self,
-        request: Request<InstantiateRequest>,
-    ) -> Result<Response<InstantiateResponse>, Status> {
-        request.handle(self).await
-    }
-
-    async fn session_create(
-        &self,
-        request: Request<SessionCreateRequest>,
-    ) -> Result<Response<SessionCreateResponse>, Status> {
-        request.handle(self).await
-    }
-
-    async fn session_set_pub_key(
-        &self,
-        request: Request<SessionSetPubKeyRequest>,
-    ) -> Result<Response<SessionSetPubKeyResponse>, Status> {
-        request.handle(self).await
     }
 }
